+++ Storage/src/main/resources/schema.sql
@@ -1,35 +1,34 @@
 CREATE TABLE IF NOT EXISTS  t_storage (
                                           id UUID PRIMARY KEY,
                          item_id UUID,
-                         quantity BIGINT,
-                         reserve BIGINT
+                         quantity INT,
+                         reserve INT
 );
-
 
 CREATE TABLE IF NOT EXISTS  car_storage (
                                             id UUID PRIMARY KEY,
                              item_id UUID,
-                             quantity BIGINT,
-                             reserve BIGINT
+                             quantity INT,
+                             reserve INT
 );
 
 CREATE TABLE IF NOT EXISTS  bike_storage (
     id UUID PRIMARY KEY,
                                             item_id UUID,
-                                            quantity BIGINT,
-                                            reserve BIGINT
+                                            quantity INT,
+                                            reserve INT
 );
 
 CREATE TABLE IF NOT EXISTS  car_wheel_storage (
                                                   id UUID PRIMARY KEY,
                                              item_id UUID,
-                                             quantity BIGINT,
-                                             reserve BIGINT
+                                             quantity INT,
+                                             reserve INT
 );
 
 CREATE TABLE IF NOT EXISTS  bike_wheel_storage (
                                                    id UUID PRIMARY KEY,
                                              item_id UUID,
-                                             quantity BIGINT,
-                                             reserve BIGINT
+                                             quantity INT,
+                                             reserve INT
 );
